fix(runtimes): share install instructions and refresh presence on WS connect
Two fixes to the Runtimes surface:

1. The /runtimes tab's setup snippet still shipped the pre-publish
   pnpm-from-repo command. Extract the welcome flow's brew/npx/direct
   channel picker into a shared `<DaemonInstallInstructions>` and
   render it in both places so the two screens stay in sync.

2. On WS reconnect the runtime stayed offline in the Runtimes tab and
   the agent detail runtime picker for up to ~30s. `RuntimeWsServer`
   only updated the in-memory hub on connect, so the DB trigger that
   fires `runtime.updated` SSE never ran and the web's React Query
   cache stayed stale until staleTime or the next HTTP heartbeat.
   Stamp `runtime.last_heartbeat` on each subscribed runtime when the
   WS upgrade completes — the existing trigger broadcasts SSE, the web
   invalidates `queryKeys.runtimes.all`, and the online dot flips.
